pyproject.toml
[project]
name = 'teletextimager'
version = '0.1'
authors = [
	{ name = 'Gavin MacGregor', email = 'gkthemac@gmail.com' },
]
description = 'Python library for rendering images of teletext pages'
readme = 'README.md'
requires-python = '>=3'
dependencies = [
	'Pillow',
	'importlib-resources>=5.0;python_version<"3.10"',
]

[build-system]
requires = [ 'setuptools' ]
build-backend = 'setuptools.build_meta'

[tool.setuptools.package-data]
teletextimager = [ 'font-etsi/*' ]

[project.scripts]
teletextimager = 'teletextimager.cli.teletextimager:main'
